Fix floating toolbar rendering.
The code to size and position the floating toolbar and it's content
when there was no overflow menu was omitted. This caused bad
rendering of the toolbar and it's content when an existing toolbar
object was updated to go in the "no overflow" state.

Fixed the issue by properly updating the floating toolbar size
and positioning of it's contents in the "no overflow" state.

Bug: 25973599
Change-Id: I9305140f493bc36b87328571cd7645f184d9cf30