Remove obsolete conditional preventing surface placement.
This conditional seems obsolete and is preventing surface changes
for dialogs on secondary displays. To show obsolesence, we can
break down our cases based on display as follows:
   1. Primary Display
    In this case we have a home stack, and we cannot enter
    this conditional.
   2. Secondary Display (no home stack)
    a. Private displays - We allow layout without a homestack
       when type == TYPE_PRIVATE_PRESENTATION
    b. Other secondary displays - As TYPE_PRIVATE_PRESENTATION
       can only be on private displays (see addWindow) we
       always enter this conditional for taskless Windows
       (e.g. Dialogs) on secondary displays, and refuse to
       apply their surface changes.
We can see if we remove the conditional, 1 will be invariant.
2a will remain invariant, except we allow layout when type !=
TYPE_PRIVATE_PRESENTATION. If there is a non private presentation
window on a private display, an invariant has malfunctioned and
we don't know what to do here (we don't enforce all the addWindow
invariants at each surface change). In case 2b we will allow surface
changes (where previously blocking them), and this seems correct,
after all we wish to allow dialogs to appear on secondary displays
right?

Bug: 26154242

Change-Id: Ia8cbb079616a80f41b131cafe46c7408a558e307