Use the right ClassLoader when restoring.
If we're restoring after a background process death, the Parcelable
creator cache is cold, and since we're handing in a null ClassLoader
the best the platform can do is try the default ClassLoader which
knows nothing about the running app.

That's why ClassLoaderCreator exists, so use it to snag the relevant
ClassLoader and pass it along.

Bug: 26075620
Change-Id: I6fd977d6178dd0f5f9c465597f5806a08097ac7c