Encode/Decode genres using CSV
Currently, the delimiter of genres is comma, but it is also used in the
broadcast genre, so we decided to encode the genre to CSV format.
Let's support only a simple CSV not a complex cases like fully double
quoted string to simplify the encode/decode method and encourage the
TIS to use these methods for broadcast genres.

Bug: 25241891
Change-Id: Id745d4fee43dbcdc508c0f6fad83477b6209ecd8