Fix bugs in logic to find a default selectd IME.
With this CL, InputMethodManagerService#resetDefaultImeLocked()
picks up the default selected IME with the same logic to find the
default enabled IMEs [1].  It should make sense because the default
selected IME should be one of the default enabled IMEs.  The previous
code is problematic because it does not check whether the IME is enabled
or not.  There was a chance that unusable IME could be picked up.

This CL also fixes the same problem to Bug 17347871 that only language
part of the locale is taken into account.

  [1] See the following series of CLs.
    - part 1: I831502db502f4073c9c2f50ce7705a4e45e2e1e3
              ed20f8d750ef0b6347448265a14ef2a2c7e1af5c
    - part 2: Ife93d909fb8a24471c425c903e2b7048826e17a3
              745e7bca8a622ffdf0d0a8e8e2485eab98182ede
    - part 3: I6571d464a46453934f0a8f5e79018a67a9a3c845
              d0dbd81fe2cd34c9a83e2f5217374d3e1a79f950
    - part 4: I871ccda787eb0f1099ba3574356c1da4b33681f3
              302df9b492d430d4b8b0ac020237f7d292cfcaeb

Bug: 27197621
Change-Id: Ia0f52c1fb9f5a68230284a1ec4829a2337b60bdd