Fixed bug with non-focusable stack activity not displaying on start.
Activity manager normally singals window manager that it is okay to
start the app transition process for an activity we are launching
during the resume process. However, the activity will not be resumed
if the activity is been launched in a non-focusable stack
(e.g. pinned stack). So, the app transtion will not start and it will
not be visible until the app transition time's out in window manager.
Activity manager now tells window manager to execute app transitions
if we are launching an activity in non-focusable stack.

Bug: 27224884
Change-Id: I35245fceb2637b927e7885b2c02daa1d55deaabb