Add more JavaDoc to InputConnectionWrapper.
This CL makes it clear that InputConnectionWrapper cannot be used to
emulate the behavior when a null InputConnection is passed to the
system.

We should provide some guideline for developers about how to deal with
bugs like crbug.com/571229, but as explained in the previous CL [1],
changing the behavior of InputConnectionWrapper could be a bit risky at
this point.  Hence we put more cautions in JavaDoc instead.

 [1] I8bc84d484ab0b27a02e74f11110430f70646e69a
     e55dd4dfec886d134e1cc56130dde6eb0ebdd1f1

Bug: 27407697
Change-Id: I98d1fc096c108603647a85bb0ba045b5dd23d37b