Fixed a bug where group expansion size calculation was wrong
On the lockscreen, the minheight was always based on the intrinsic
height and not actually the minHeight of the children. Because
children were userlocked, they were now reporting their small size
instead of the single line size.
This meant that the expand motion on the lockscreen wasn't following
the finger at all.

Bug: 27418617
Change-Id: I81daaf87887de1adc014cb4c6d46f4ef118061e5