Don't clear effects when the panel is fully collapsed.
Otherwise, a FLAG_SHOW_WHEN_LOCKED activity would be unable
to make noise (particularly in conjunction with a long
sound, or FLAG_INSISTENT) because the mere presence of the
status bar on screen would generate an updateVisibleToUser()
call, which would in turn use onPanelRevealed() to call
clearEffects().

Instead, we have to clear effects from deep in the stack
scroller; that's apparently the only place right now where
we know with 100% certainty that notifications are actually
visible to the user.

We also no longer clear effects just because the keyguard is
shown. That's silly.

Bug: 26863782
Change-Id: I9ec53024f4be060da203b835d31da81104e8c8e6