Fix extra data in cache
Reverting the revert.
The original CL(commit c2e851f36a5a6d35584e0baed204560c076df984) had a 
bug in it. It was calling accountExistsCacheLocked(), while holding 
accounts.cacheLock. The function accountExistsCacheLocked, was in turn
calling into getUserAccounts, which acquires mUsers. And this causes
trouble. mUsers is a a lock on all accounts, and hence calling it after
holding accounts.cachelock is calling for trouble. Since the locks are 
acquired in other order, it causes a potential deadlock issue, which we 
discovered later on.

That bug was fixed by commit eec7542220cdb965f1e713d531934b0842cb23b6.
We already have useraccount object and should reuse it.

This reverts commit 381a65af064348b1346016157db74c943205e217. This 
commit was unneeded as the bug had been fixed by that time.

Change-Id: I5328c31fd485fd2c1c652cd0e7c2c4bded38a5fd