Disable suppressLayout in Visibility Activity Transitoins.
Bug 27389255

In Idf21542464a13bac7b4d4a17f6b9303f68d550c3, I had removed
a suppressLayout check from forced visibility. Unfortunately,
when the transition ends, it calls layout if there was a
requestLayout during the transition. This prevents the
suppression at the level of the individual views so that it
can be handled at the decor View level as it was before.

Change-Id: I0f016e6356fd1ceff705b122a6b4ac3f3f09600d